camera/pyproject.toml
[tool.poetry]
name = "camera"
version = "0.1.0"
description = ""
authors = ["yuichiroaoki <45054071+yuichiroaoki@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
ruff = "^0.0.282"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
